typing-10 修改时间

+++ src/SwingModule.java
@@ -71,7 +71,7 @@ class SwingModule {
 
 	int delCount;
 
-	long numberOfWords;
+	float numberOfWords;
 
 	Date startTime = new Date();
